リファクタリング: 全体的なリファクタリングとドキュメント改善
実施した内容は以下の通りです。

- **ソースコードのリファクタリング:**
  - `yamap_auto_domo.py` にあった非アクティブユーザーのアンフォロー機能を `unfollow_utils.py` に分離し、責務を明確化しました。
  - 全体的なコードの可読性を向上させるための軽微な修正を行いました。

- **ドキュメントの改善:**
  - `README.md` を全面的に見直し、機能一覧、セットアップ方法、設定ファイルの説明などをより分かりやすく更新しました。特に、並列処理のリスクに関する注意喚起を追記しました。
  - `TASK_MANAGEMENT.md` をクリアし、今後の課題として「Google Cloud Functionsでの定期実行」を明記しました。

- **Cloud Functions対応準備:**
  - Cloud Functionsの実行環境を想定し、`requirements.txt` の内容を再確認しました。
  - デプロイの際にエントリーポイントとして使用する `main.py` を新規に作成しました。
